lectures/demos/neural_network_demo/requirements.txt
streamlit==1.24.0
numpy==1.24.3
matplotlib==3.7.1
pandas==2.0.2
scikit-learn==1.2.2
Pillow==9.5.0 